bo sung api cap nhat thoi gian hoc tap cho nguoi dung

+++ enrollment-service/src/main/java/com/pbl/elearning/enrollment/payload/response/EnrollmentResponse.java
@@ -21,6 +21,6 @@ public class EnrollmentResponse {
     private Timestamp completionDate;
     private Double progressPercentage;
     private EnrollmentStatus status;
-    private Integer totalWatchTimeMinutes;
+    private Double totalWatchTimeMinutes;
     private Timestamp lastAccessedAt;
 }
